pyproject.toml
[project]
name = "paperless-mcp-server"
version = "0.1.0"
description = "A read-only MCP server for Agentic RAG over Paperless-ngx documents."
readme = "README.md"
authors = [
    { name = "Dr. Henning Dickten" }
]
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "mcp>=0.1.0",
    "httpx>=0.27.0",
    "chromadb>=1.5.2",
    "google-genai>=0.5.0",
    "tenacity>=8.2.3",
    "loguru>=0.7.2",
    "pydantic-settings>=2.2.1"
]

dev = [
    "pytest>=8.1.1",
    "ruff>=0.3.4",
    "pytest-asyncio>=0.23.6",
    "respx>=0.21.1",
    "pytest-env>=1.1.3"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "api",
    "core",
    "semantic",
    "server"
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
env = [
    "PAPERLESS_URL=http://mock",
    "PAPERLESS_TOKEN=mock",
    "GEMINI_API_KEY=mock"
]
